pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Flask backend for an AI RAG chatbot using LangChain, HuggingFce and Pinecone."
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "flask",
    "flask-cors",
    "python-dotenv",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.1.0",
    "langchain-pinecone>=0.1.0",
    "langchain-huggingface>=0.1.0",
    "pinecone-client>=4.1.0",
    "huggingface-hub>=0.25.0",
    "sentence-transformers>=3.0.0",
]